test: move specs/basicAuth to ESM
